Favorites Context, Home y Register
+ Emparejamiento de receta por usuario para que se asignen de manera individual por cada usuario respectivamente
/ Correccion de estilos y vibracion
+ Se agrego que ahora en el home se muestre el nombre del usuario loggeado

+++ src/context/FavoriteContext.js
@@ -1,10 +1,11 @@
 import React, { createContext, useState, useEffect, useContext } from "react";
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { initializeApp } from "firebase/app";
-import { getFirestore, collection, getDocs, doc, addDoc, deleteDoc, setDoc } from "firebase/firestore";
+import { getFirestore, collection, getDocs, doc, addDoc, deleteDoc, setDoc, getDoc } from "firebase/firestore";
 
 //Credenciales a Firebase
-import app from '../components/FirebaseConfig';
+import { app } from '../components/FirebaseConfig';
+import { onAuthStateChanged, getAuth } from "firebase/auth";
 
 const FavoritesContext = createContext();
 
@@ -24,26 +25,64 @@ const database = getFirestore(app);
 
 export const FavoritesProvider = ({ children }) => {
     const [favorites, setFavorites] = useState([]);
+    const [user, setUser] = useState(null);
 
     useEffect(() => {
-        loadFavorites();
+        const auth = getAuth();
+        const unsubscribe = onAuthStateChanged(auth, async (currentUser) => {
+            setUser(currentUser);
+
+            if (currentUser) {
+                try {
+                    const userRef = doc(database, "users", currentUser.uid);
+                    const userSnap = await getDoc(userRef);
+
+                    if (!userSnap.exists()) {
+                        await setDoc( userRef, {
+                            name: currentUser.displayName || "Cocinero",
+                            email: currentUser.email,
+                        });
+                        console.log("Usuario registrado en Firestore: ", currentUser.displayName);
+                    }
+
+                    loadFavorites(currentUser.uid);
+                } catch (error) {
+                    console.log("Error al registrar usuario en Firestore: ", error.message);
+                }
+            } else {
+                setUser(null);
+                setFavorites([]); // Si el usuario cierra sesión, limpia los favoritos
+            }
+        });
+
+        return () => unsubscribe();
+
     }, []);
 
-    const loadFavorites = async () => {
+    const loadFavorites = async (userId) => {
         try {
             
-            // Cargar desde AsyncStorage
-            const storedFavorites = await AsyncStorage.getItem("favorites");
-            if (storedFavorites) {
-                setFavorites(JSON.parse(storedFavorites));
+            if (!userId) {
+                return;
             }
 
-            // Cargar desde Firestore
-            const favoritesCollection = collection(database, "favorites");
-            const snapshot = await getDocs(favoritesCollection);
-            
+            // Obtener el nombre del usuario desde Firestore
+            const userDocRef = doc(database, "users", userId);
+            const userSnap = await getDoc(userDocRef);
+
+            let userName = "Usuario desconocido";
+
+            if (userSnap.exists()) {
+                userName = userSnap.data().name || "Usuario sin nombre";
+            }
+
+            // Obtener favoritos
+            const userFavoritesCollection = collection(database, "users", userId, "favorites");
+            const snapshot = await getDocs(userFavoritesCollection);
+
             const favoritesList = snapshot.docs.map((doc) => ({
                 idMeal: doc.id,
+                name: doc.data().name,
                 ...doc.data(),
             }));
 
@@ -67,11 +106,17 @@ export const FavoritesProvider = ({ children }) => {
     };
 
     const addFavorite = async (meal) => {
+
+        if (!user) {
+            console.log("El usuario no ha iniciado sesion.");
+            return;
+        }
+
         try {
-            const favoritesCollection = collection(database, "favorites");
+            const userfavoritesCollection = collection(database, "users", user.uid, "favorites");
 
             // Guarda la receta con su idMeal como ID en Firestore
-            const favoriteDoc = doc(favoritesCollection, meal.idMeal.toString()); //Agrega la receta a la coleccion de favoritos en Firestore
+            const favoriteDoc = doc(userfavoritesCollection, meal.idMeal.toString()); //Agrega la receta a la coleccion de favoritos en Firestore
             await setDoc(favoriteDoc, meal);
 
             // Actualizar estado y guardar en AsyncStorage
@@ -89,12 +134,17 @@ export const FavoritesProvider = ({ children }) => {
     };
 
     const removeFavorite = async (mealId) => {
+
+        if (!user) {
+            console.log("El usuario no ha iniciado sesion.");
+            return;
+        }
         
         try {
 
             console.log("Intentando eliminar: ", mealId); // Verificar qué ID se está enviando
 
-            const favoriteDoc = doc(database, "favorites", mealId.toString()); // Convertir ID a string
+            const favoriteDoc = doc(database, "users", user.uid, "favorites", mealId.toString()); // Convertir ID a string
             
             await deleteDoc(favoriteDoc);
             console.log("Receta eliminada en Firestore");
@@ -111,11 +161,10 @@ export const FavoritesProvider = ({ children }) => {
         } catch (error) {
             console.error("Erro al eliminar la receta favorita ", error);
         }
-
     };
 
     return (
-        <FavoritesContext.Provider value={{ favorites, addFavorite, removeFavorite }}>
+        <FavoritesContext.Provider value={{ favorites, user, addFavorite, removeFavorite }}>
             {children}
         </FavoritesContext.Provider>
     );
